Moved helpers to a folder, changed dla_3dsim and added header comments to the files
